requirements.txt
streamlit==1.50.0
chromadb==1.5.0
python-dotenv==1.2.1
llama-index-core
llama-index-embeddings-openai
llama-index-llms-openai
llama-index-vector-stores-chroma
openai==1.109.1
